requirements.txt
asgiref             3.5.1
Django              4.0.4
django-filter       21.1
djangorestframework 3.13.1
Pillow              9.1.0
pip                 22.0.2
psycopg2            2.9.3
pytz                2022.1
setuptools          59.6.0
sqlparse            0.4.2
wheel               0.37.1